# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pocket Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AgentStatus = typeof AgentStatus[keyof typeof AgentStatus];


export const AgentStatus = {
  active: 'active',
  inactive: 'inactive',
  paused: 'paused',
} as const;

export interface Agent {
  id: string;
  name: string;
  description?: string;
  status: AgentStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAgentInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /** @maxLength 500 */
  description?: string;
  status?: AgentStatus;
}

export interface UpdateAgentInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name?: string;
  /** @maxLength 500 */
  description?: string;
  status?: AgentStatus;
}

export interface ErrorResponse {
  error: string;
  message: string;
}