# lib/api-zod/src/generated/types/agent.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pocket Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentStatus } from './agentStatus';

export interface Agent {
  id: string;
  name: string;
  description?: string;
  status: AgentStatus;
  createdAt: Date;
  updatedAt: Date;
}