# lib/api-zod/src/generated/types/agentStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pocket Agent API
 * OpenAPI spec version: 0.1.0
 */

export type AgentStatus = typeof AgentStatus[keyof typeof AgentStatus];


export const AgentStatus = {
  active: 'active',
  inactive: 'inactive',
  paused: 'paused',
} as const;