# lib/api-zod/src/generated/types/updateAgentInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pocket Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentStatus } from './agentStatus';

export interface UpdateAgentInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name?: string;
  /** @maxLength 500 */
  description?: string;
  status?: AgentStatus;
}